pyproject.toml
[project]
name = "pokemon-natural-geo"
version = "0.1.0"
description = "Photorealistic Pokémon nature documentary production pipeline"
requires-python = ">=3.10"
dependencies = [
    "google-generativeai>=0.8.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "pyjwt>=2.8.0",
    "requests>=2.31.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "openai>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
telegram = ["python-telegram-bot>=21.0"]

[project.scripts]
pipeline-server = "backend.app:main"

[tool.uv]
dev-dependencies = []
